refactor: useFunnelStore를 전역 store 디렉토리로 이동
- generate·home·analytics 피처가 imageSetup 피처의 useFunnelStore를 직접 import하던 cross-feature 위반 해소
- S2(다음 작업)에서 state machine 전이 처리 함수(completeGenerationRequest/abandonFlow)가 퍼널 reset을 합법적으로 소유하기 위한 전제
- 파일 이동 + import 경로 10곳 갱신, 동작 변화 없음
