+++ src/pages/imageSetup/hooks/activityInfo/useActivityInfo.ts
@@ -6,6 +6,7 @@ import { trackSelectFurnitureBtnCtaClick } from '@pages/imageSetup/analytics/sel
 
 import { ROUTES } from '@routes/paths';
 
+import { useFunnelStore } from '@store/useFunnelStore';
 import { holdEntryRoute } from '@store/useImageFlowStore';
 
 import {
@@ -20,7 +21,6 @@ import { useCategorySelection } from './useCategorySelection';
 import { useGlobalConstraints } from './useGlobalConstraints';
 import { useActivitiesQuery } from '../../apis/queries/useActivitiesQuery';
 import { useFurnitureCategoriesQuery } from '../../apis/queries/useFurnitureCategoriesQuery';
-import { useFunnelStore } from '../../stores/useFunnelStore';
 import { CATEGORY_SELECTION_MODE } from '../../types/funnel/activityInfo';
 import { useRecentFloorPlanQuery } from '../../v2/apis/queries/useRecentFloorPlanQuery';
